backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
numpy
pandas
scikit-learn
faiss-cpu
requests
